# src/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  request.RefreshTokens:
    properties:
      refresh_token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VybmFtZSI6InByb2YiL...
        type: string
    type: object
  request.SignIn:
    properties:
      email:
        example: example@email.com
        type: string
      password:
        example: luc@ssantos123
        type: string
    type: object
  request.SignUp:
    properties:
      email:
        example: example@email.com
        type: string
      first_name:
        example: Lucas
        type: string
      last_name:
        example: Santos
        type: string
      password:
        example: luc@ssantos123
        type: string
    type: object
...
  swagger.HttpError:
    properties:
      message:
        example: Something went wrong...
        type: string
    type: object
...
  swagger.User:
    properties:
      access_token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VybmFtZSI6InByb2YiL...
        type: string
      created_at:
        example: "2024-01-06T14:52:23.723959Z"
        type: string
      email:
        example: example@email.com
        type: string
      first_name:
        example: Lucas
        type: string
      id:
        example: 54b094b23-f426-89d2-843e-1335410b18df
        type: string
      last_name:
        example: Santos
        type: string
      refresh_token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VybmFtZSI6InByb2YiL...
        type: string
      roles:
        example:
        - user
        items:
          type: string
        type: array
      updated_at:
        example: "2024-01-06T16:02:29.854799Z"
        type: string
    type: object
...
info:
  contact:
    email: email.vocenacamara@gmail.com
    name: Você na Câmara
  description: Set of endpoints that make up the backend of the Você na Câmara platform,
    structured to enable communication between the services and the execution of the
    system's functionalities.
  title: Você na Câmara API
  version: v1
paths:
...
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: This request is responsible for refreshing the user's access tokens
        on the platform.
      operationId: Refresh
      parameters:
      - description: Request body
        in: body
        name: requestBody
        required: true
        schema:
          $ref: '#/definitions/request.RefreshTokens'
      produces:
      - application/json
      responses:
        "200":
          description: Successful request
          schema:
            $ref: '#/definitions/swagger.User'
        "400":
          description: Badly formatted request
          schema:
            $ref: '#/definitions/swagger.HttpError'
        "401":
          description: Unauthorized access
          schema:
            $ref: '#/definitions/swagger.HttpError'
        "422":
          description: Some of the data provided is invalid
          schema:
            $ref: '#/definitions/swagger.HttpError'
        "500":
          description: An unexpected error occurred while processing the request
          schema:
            $ref: '#/definitions/swagger.HttpError'
        "503":
          description: Some of the services/resources are temporarily unavailable
          schema:
            $ref: '#/definitions/swagger.HttpError'
      summary: Refresh access tokens
      tags:
      - Authentication
  /auth/sign-in:
    post:
      consumes:
      - application/json
      description: This request is responsible for signing the user into the platform.
      operationId: SignIn
      parameters:
      - description: Request body
        in: body
        name: requestBody
        required: true
        schema:
          $ref: '#/definitions/request.SignIn'
      produces:
      - application/json
      responses:
        "200":
          description: Successful request
          schema:
            $ref: '#/definitions/swagger.User'
        "400":
          description: Badly formatted request
          schema:
            $ref: '#/definitions/swagger.HttpError'
        "401":
          description: Unauthorized access
          schema:
            $ref: '#/definitions/swagger.HttpError'
        "422":
          description: Some of the data provided is invalid
          schema:
            $ref: '#/definitions/swagger.HttpError'
        "500":
          description: An unexpected error occurred while processing the request
          schema:
            $ref: '#/definitions/swagger.HttpError'
        "503":
          description: Some of the services/resources are temporarily unavailable
          schema:
            $ref: '#/definitions/swagger.HttpError'
      summary: Sign In
      tags:
      - Authentication
  /auth/sign-out:
    post:
      description: This request is responsible for signing the user out of the platform.
      operationId: SignOut
      produces:
      - application/json
      responses:
        "204":
          description: Successful request
        "400":
          description: Badly formatted request
          schema:
            $ref: '#/definitions/swagger.HttpError'
        "401":
          description: Unauthorized access
          schema:
            $ref: '#/definitions/swagger.HttpError'
        "403":
          description: Access denied
          schema:
            $ref: '#/definitions/swagger.HttpError'
        "500":
          description: An unexpected error occurred while processing the request
          schema:
            $ref: '#/definitions/swagger.HttpError'
        "503":
          description: Some of the services/resources are temporarily unavailable
          schema:
            $ref: '#/definitions/swagger.HttpError'
      security:
      - BearerAuth: []
      summary: Sign Out
      tags:
      - Authentication
  /auth/sign-up:
    post:
      consumes:
      - application/json
      description: This request is responsible for signing the user up to the platform.
      operationId: SignUp
      parameters:
      - description: Request body
        in: body
        name: requestBody
        required: true
        schema:
          $ref: '#/definitions/request.SignUp'
      produces:
      - application/json
      responses:
        "201":
          description: Successful request
          schema:
            $ref: '#/definitions/swagger.User'
        "400":
          description: Badly formatted request
          schema:
            $ref: '#/definitions/swagger.HttpError'
        "401":
          description: Unauthorized access
          schema:
            $ref: '#/definitions/swagger.HttpError'
        "409":
          description: Some of the data provided is conflicting
          schema:
            $ref: '#/definitions/swagger.HttpError'
        "422":
          description: Some of the data provided is invalid
          schema:
            $ref: '#/definitions/swagger.HttpError'
        "500":
          description: An unexpected error occurred while processing the request
          schema:
            $ref: '#/definitions/swagger.HttpError'
        "503":
          description: Some of the services/resources are temporarily unavailable
          schema:
            $ref: '#/definitions/swagger.HttpError'
      summary: Sign Up
      tags:
      - Authentication
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"